Add files via upload
Stuff done:
I check for name, class and image (it must be chosen and entered).
Saves the data to the core data (please check this, but it should work), and also sets the initial level to 1.
Highlights the selected image in blue (can be changed)

Errors:
Can't figure out how to remove the blue border from the selected image when the data is saved (maybe let's ask the professor).

There's something wrong with the way the images are displayed in the collectionView, but i can't seem to find the error, can you guys try and fix it? (i've made the background green so we can see that the images are really there. 

Notes:
Remember to implement the update of the tableView when a new adventurer is added so it updates right when the new adventurer is added or when the user returns to the tableView

The adventurer's hit points and attack power can be set in the function "save" (//Function to save data to CoreData) if you want to do it here.